+++ src/main/java/net/frontlinesms/data/ConfigurableService.java
@@ -1,0 +1,10 @@
+package net.frontlinesms.data;
+
+import net.frontlinesms.data.domain.PersistedSettings;
+
+public interface ConfigurableService {
+	StructuredProperties getPropertiesStructure();
+	PersistedSettings getSettings();
+	void setSettings(PersistedSettings settings);
+	Class<? extends ConfigurableService> getSuperType();
+}
